feat: upgrade go dependencies
This commit upgrades the go dependencies to their latest versions.
The gosec linter was disabled for some lines because they are not
security issues.
The expire package was removed because it was not being used.
The expire_test package was moved to beacon_test package.
The CalculateSlotExpiration and GetSlotTime functions were moved to
the expire_test package.