feat: Electra support (#177)
* feat: Initial Electra support

* feat: Update Go version to 1.22.0

* style: Update Go version in go.mod to 1.22

* feat: Update Go versions to 1.22 in workflow files

* fix: Update golang version in Dockerfile

* Bump go-eth2

* fix(web): handle epoch as string

* refactor: Update dependencies versions and modules in go.mod and go.sum files

* refactor: Simplify fetching current epoch in download functions

* refactor: Add NotOptimisticEL method to Nodes type

* chore(goeth2): bump to latest electra commit

* fix(beacon): deneb fork name

* feat: upgrade go dependencies

This commit upgrades the go dependencies to their latest versions.
The gosec linter was disabled for some lines because they are not
security issues.
The expire package was removed because it was not being used.
The expire_test package was moved to beacon_test package.
The CalculateSlotExpiration and GetSlotTime functions were moved to
the expire_test package.

* fix(beacon): remove nolint directive from UpstreamsStatus function

* feat(checkpoint-sync): upgrade upload-artifact action to v4

The action was upgraded to the latest major version.

* feat(checkpoint-sync): use secrets for beacon node URLs

* feat(checkpoint-sync): use env vars for beacon node URLs in action

* feat(checkpoint-sync): add beacon_node_url input to action

This allows the user to specify the beacon node URL to use for
checkpoint sync, and removes the need for the action to have
knowledge of the different networks and their corresponding URLs.
This makes the action more generic and reusable.

* feat(integration.yaml): use env vars for beacon node URLs
feat: remove exportloopref linter

The integration workflow now uses environment variables to
configure beacon node URLs, and the exportloopref linter has been
removed from the golangci configuration.

* feat: strip trailing slash from beacon node URL in checkpoint sync
fix(integration.yaml): set beacon_node_url to MAINNET_BEACON_API_URL

* fix(checkpoint-sync): validate beacon node URL format

The action now validates the beacon node URL to ensure it starts
with either 'http://' or 'https://'. If the URL is invalid, the
action will exit with an error message.

* feat(integration.yaml): add mainnet integration tests and use secrets for beacon node url

* feat(action.yaml): add execution endpoint to lighthouse to connect to geth

---------

Co-authored-by: Andrew Davis <1709934+Savid@users.noreply.github.com>
Co-authored-by: Matty Evans <mattevansnz@gmail.com>