feat: add support for fulu fork
This commit adds support for the "fulu" fork to the checkpointz.
It includes changes to the go.mod and go.sum files to update
dependencies, as well as modifications to the API handler and web
interface to handle the new fork. The API handler now includes a case
for "fulu" when determining the content type resolvers for beacon
blocks and debug beacon states. The web interface has been updated to
include a case for "FULU" when retrieving block data for a specific
slot. Additionally, the APIBeaconBlock type has been updated to
include "FULU" as a possible value for the Version property, as well
as a Fulu property for storing the block data.