feat: Fulu (#206)
* feat: add support for fulu fork

This commit adds support for the "fulu" fork to the checkpointz.
It includes changes to the go.mod and go.sum files to update
dependencies, as well as modifications to the API handler and web
interface to handle the new fork. The API handler now includes a case
for "fulu" when determining the content type resolvers for beacon
blocks and debug beacon states. The web interface has been updated to
include a case for "FULU" when retrieving block data for a specific
slot. Additionally, the APIBeaconBlock type has been updated to
include "FULU" as a possible value for the Version property, as well
as a Fulu property for storing the block data.

* build: update go dependencies to latest versions

* build: update github actions workflows to use ubuntu-24

* ci: update ubuntu version in workflows to 24.04

* fix(api): correct block and state version handling for Fulu in handlers

* build: update go-eth2-client dependency to latest commit

* fix: update go-eth2-client dependency and use spec.DataVersionFulu constant

* chore: update dependencies to latest versions

This commit updates the following dependencies:
- github.com/attestantio/go-eth2-client from v0.25.0 to v0.27.1
- github.com/ethpandaops/beacon from v0.51.0 to v0.64.0

The `replace` directive for `github.com/attestantio/go-eth2-client` has been removed as it is no longer necessary with the updated version.

These updates are made to incorporate the latest features and bug fixes from the respective libraries.

---------

Co-authored-by: Barnabas Busa <busa.barnabas@gmail.com>